% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesMig.R
\docType{package}
\name{bayesMig-package}
\alias{bayesMig-package}
\title{Bayesian Projection of Migration}
\description{
Collection of functions for making probabilistic projections of net migration rate for all countries of the world,
using a Bayesian hierarchical model (BHM) and the United Nations demographic time series.
Methodological details are provided in Azose & Raftery (2015)
}
\details{
_bayesMig
}
\examples{
\dontrun{
#Run simulations and write to file
run.mig.mcmc(nr.chains=4, iter=10000, thin=10, replace.output = TRUE)
mig.diagnose(sim.dir='./bayesMig.output', burnin=100, thin=1)

#Predict and plot
mig.pred=mig.predict(burnin=100, replace.output = TRUE)
mig.trajectories.plot.all(mig.pred)
}

}
\references{
Azose, J. J., & Raftery, A. E. (2015). 
Bayesian probabilistic projection of international migration. Demography, 52(5), 1627-1650.
}
\author{
Jon Azose, Hana Sevcikova and Adrian Raftery

Maintainer: Hana Sevcikova <hanas@uw.edu>
}
