% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outputs.R
\name{coda.list.mcmc}
\alias{coda.list.mcmc}
\title{Conversion to coda-formatted objects}
\usage{
coda.list.mcmc(mcmc=NULL, country=NULL, chain.ids=NULL, 
sim.dir=file.path(getwd(), 'bayesMig.output'),
par.names=mig.parameter.names(), 
par.names.cs=mig.parameter.names.cs(), 
rm.const.pars=FALSE, burnin=0, ...)
}
\arguments{
\item{mcmc}{A list of objects of class \code{bayesMig.mcmc}. If \code{NULL}, the MCMCs are
loaded from \code{sim.dir}. Either \code{mcmc} or \code{sim.dir} must be given.}

\item{country}{Country name or code. Used in connection with the \code{par.names.cs} argument
(see below)}

\item{chain.ids}{Vector of chain identifiers. By default, all chains available in the \code{mcmc.list}
object are included}

\item{sim.dir}{Directory with the MCMC simulation results. Only used if \code{mcmc} is \code{NULL}}

\item{par.names}{Names of country-independent parameters to be included. Default names are
those returned by the \code{mig.parameter.names} function, which includes all country-independent
parameters in the BHM.}

\item{par.names.cs}{Names of country-specific parameters to be included. The argument \code{country}
is used to filter out traces that correspond to a specific country. If \code{country} is not given, 
traces of each parameter are given for all countries. Default names are those returned by 
\code{mig.parameter.names.cs()}, which includes all country-specific parameters in the BHM.}

\item{rm.const.pars}{Logical indicating if parameters with constant values should be removed.}

\item{burnin}{Number of iterations that should be removed from the beginning of each chain.}

\item{...}{Other variables passed to called functions}
}
\description{
Conversion to coda-formatted objects
}
