% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryCodeVec_bigCountries.R
\docType{data}
\name{countryCodeVec_bigCountries}
\alias{countryCodeVec_bigCountries}
\title{List of default country codes}
\format{An object of class \code{numeric} of length 201.}
\usage{
countryCodeVec_bigCountries
}
\description{
Default list of country codes for inclusion in analysis.
This is the countries in the WPP2017 dataset with population of at least 100,000.
}
\keyword{datasets}
