% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outputs.R
\name{get.mig.parameter.traces}
\alias{get.mig.parameter.traces}
\alias{get.mig.parameter.traces.cs}
\title{Accessing MCMC Parameter Traces}
\usage{
get.mig.parameter.traces(
  mcmc.list,
  par.names = NULL,
  burnin = 0,
  thinning.index = NULL,
  thin = NULL
)

get.mig.parameter.traces.cs(
  mcmc.list,
  country.obj,
  par.names = NULL,
  burnin = 0,
  thinning.index = NULL,
  thin = NULL
)
}
\arguments{
\item{mcmc.list}{List of \code{\link{bayesMig.mcmc}} objects.}

\item{par.names}{Names of country-independent parameters (in case of 
\code{get.mig.parameter.traces}) or country-specific parameters 
(in case of \code{get.mig.parameter.traces.cs}) to be included. 
By default all parameters are included.}

\item{burnin}{Burn-in indicating how many iterations should be removed 
from the beginning of each chain.}

\item{thinning.index}{Index of the traces for thinning. If it is \code{NULL}, 
\code{thin} is used. \code{thinning.index} does not include \code{burnin} 
and should be flattened over all chains. For example, if there are two MCMC 
chains of length 1000, \code{burnin=200} and we want an equidistantly spaced 
sample of length 400, then the value should be 
\code{thinning.index=seq(1,1600, length=400)}.}

\item{thin}{An integer value for thinning. It is an alternative to 
\code{thinning.index}. The above example is equivalent to \code{thin=4}.}

\item{country.obj}{Country object (see \code{\link[bayesTFR]{get.country.object}}).}
}
\value{
Both functions return a matrix with columns being the parameters and 
    rows being the MCMC values, attached to one another in case of multiple chains. 
    \code{get.mig.parameter.traces} returns country-independent parameters, 
    \code{get.mig.parameter.traces.cs} returns country-specific parameters.
}
\description{
Functions for accessing traces of the MCMC parameters, either country-independent 
    or country-specific.
}
\seealso{
\code{\link{mig.coda.list.mcmc}} for another way of retrieving parameter traces.
}
