% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outputs.R
\name{get.mig.mcmc}
\alias{get.mig.mcmc}
\title{Access MCMC results}
\usage{
get.mig.mcmc(
  sim.dir = file.path(getwd(), "bayesMig.output"),
  chain.ids = NULL,
  low.memory = TRUE,
  burnin = 0,
  verbose = FALSE
)
}
\arguments{
\item{sim.dir}{Directory where simulation results are stored}

\item{chain.ids}{Chain identifiers in case only specific chains should be included
in the resulting object. By default, all available chains are included.}

\item{low.memory}{Logical. If \code{FALSE} full MCMC traces are loaded into memory.}

\item{burnin}{Burn-in used for loading traces. Only relevant, if \code{low.memory=FALSE}.}

\item{verbose}{Logical value. Switches log messages on and off.}
}
\description{
This function retrieves results of an MCMC simulation and creates an object of class
\code{bayesMig.mcmc.set}.
}
