% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outputs.R
\name{get.mig.trajectories}
\alias{get.mig.trajectories}
\title{Accessing Trajectories of Net Migration Rate}
\usage{
get.mig.trajectories(mig.pred, country)
}
\arguments{
\item{mig.pred}{Object of class \code{\link{bayesMig.prediction}}.}

\item{country}{Name or numerical code of a country. It can also be given as
ISO-2 or ISO-3 characters.}
}
\value{
Array of size the number of projection periods (including the present year) times the number of trajectories.
}
\description{
Function for accessing all future trajectories of the net migration rate 
    from a prediction object in a form of an array.
}
\details{
The function loads projected trajectories of net migration rate for the given country from disk
    and returns it as a matrix.
}
\seealso{
\code{\link{bayesMig.prediction}}, \code{\link{get.mig.prediction}}, \code{\link{mig.trajectories.table}}
}
