% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_storage.R
\name{store.bayesMig.convergence}
\alias{store.bayesMig.convergence}
\alias{store.bayesMig.prediction}
\title{Internal Functions and datasets of bayesMig}
\usage{
store.bayesMig.convergence(diag, thin, burnin, output.dir)

store.bayesMig.prediction(pred, output.dir = NULL)
}
\value{
None
}
\description{
These functions and datasets are not to be used directly by the user.
}
\keyword{internal}
