% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_objects.R
\name{mcmc-meta-object}
\alias{mcmc-meta-object}
\alias{bayesMig.mcmc.meta}
\title{MCMC Meta Object}
\value{
Stores values of the various input arguments of the {\code{\link{run.mig.mcmc}}} function.
These are \code{nr.chains}, \code{start.year}, \code{present.year}, \code{wpp.year}, \code{my.e0.file}
}
\description{
Meta object \code{bayesMig.mcmc.meta} used by all chains of the same MCMC simulation. 
It contains information that is common to all chains. It is a part of a {\code{\link{bayesMig.mcmc.set}}} object.
}
\details{
The object is in standard cases not to be manipulated by itself.
}
