% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_objects.R
\name{mcmc-object}
\alias{mcmc-object}
\alias{bayesMig.mcmc}
\title{MCMC Object}
\description{
MCMC object \code{bayesMig.mcmc} containing information about one MCMC chain. 
A set of such objects belonging to the same simulation, together with a {\code{\link{bayesMig.mcmc.meta}}} object, 
constitute a {\code{\link{bayesMig.mcmc.set}}} object.
}
