% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{mig.trajectories.plot.all}
\alias{mig.trajectories.plot.all}
\title{Graphical output of posterior distribution of migration trajectories}
\usage{
mig.trajectories.plot.all(mig.pred, output.dir=file.path(getwd(), 'migTrajectories'),
output.type="png", main=NULL, verbose=FALSE, ...)
}
\arguments{
\item{mig.pred}{Prediction object of class \code{bayesMig.prediction}}

\item{output.dir}{Directory into which resulting plots are written}

\item{output.type}{Type of the resulting plot files. Can be "png", "pdf", "jpeg", "bmp",
"tiff", or "postscript"}

\item{main}{Main title for the plots. Any occurrence of the string "XXX" is replaced by the
name of the appropriate country.}

\item{verbose}{Logical value. Switches log messages on and off.}

\item{...}{Other variables passed to called functions}
}
\description{
This function plots the posterior distribution of migration trajectories for all countries,
including their median and given probability intervals
}
