% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mig.R
\name{mig.write.projection.summary}
\alias{mig.write.projection.summary}
\title{Writing Projection Summary Files}
\usage{
mig.write.projection.summary(pred, output.dir, ...)
}
\arguments{
\item{pred}{Object of class \code{bayesMig.prediction}.}

\item{output.dir}{Directory where output is written.}

\item{\dots}{Additional arguments passed to the underlying functions. 
Here, argument \code{precision} can be set to determine the number 
of significant digits (default is 4).}
}
\value{
No return value.
}
\description{
The function creates two files containing projection summaries, 
    such as the median, the lower and upper bound of the 80 and 90\% probability 
    intervals, respectively, and the constant variant. One file is in a user-friendly 
    format, whereas the other is in a UN-specific format with internal coding of the 
    time and the variants.
}
\seealso{
\code{\link[bayesTFR]{write.projection.summary}}
}
