% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outputs.R
\name{mig.parameter.names}
\alias{mig.parameter.names}
\alias{mig.parameter.names.cs}
\title{Accessing Parameter Names}
\usage{
mig.parameter.names()

mig.parameter.names.cs(country.code = NULL)
}
\arguments{
\item{country.code}{Location code. If it is given, the country-specific parameter names contain 
the suffix \sQuote{_country\eqn{X}} where \eqn{X} is the \code{country.code}.}
}
\value{
\code{mig.parameter.names} returns names of the world parameters.

\code{mig.parameter.names.cs} returns names of the country-specific parameters.
}
\description{
Functions for accessing names of the MCMC parameters, 
    either country-independent or country-specific.
}
\examples{
mig.parameter.names()

mig.parameter.names.cs()

}
