% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcmc.R
\name{run.mig.mcmc}
\alias{run.mig.mcmc}
\title{Run Markov chain Monte Carlo for parameters of net migration rate model}
\usage{
run.mig.mcmc(
  nr.chains = 3,
  iter = 50000,
  output.dir = file.path(getwd(), "bayesMig.output"),
  thin = 1,
  replace.output = FALSE,
  start.year = 1950,
  present.year = 2020,
  wpp.year = 2019,
  my.mig.file = NULL,
  sigma.c.min = 1e-04,
  a.up = 10,
  a.ini = NULL,
  a.half.width = 0.3,
  mu.range = c(-0.5, 0.5),
  sigma.mu.range = c(0, 0.5),
  mu.ini = NULL,
  exclude.from.world = NULL,
  seed = NULL,
  parallel = FALSE,
  nr.nodes = nr.chains,
  buffer.size = 1000,
  verbose = TRUE,
  verbose.iter = 10,
  ...
)
}
\arguments{
\item{nr.chains}{An integer number of independent Markov chains to run.}

\item{iter}{The number of iterations to run per Markov chain.}

\item{output.dir}{A file path pointing to the directory in which to store results.}

\item{thin}{Thinning interval -- A chain with 1000 iterations thinned by 20 will return a 
final count of 50 iterations.}

\item{replace.output}{If the specified output directory already exists, should it be overwritten?}

\item{start.year}{Start year for using historical data.}

\item{present.year}{End year for using historical data.}

\item{wpp.year}{Year for which WPP data is used if no user data is provided via \code{my.mig.file}. 
In such a case, the function loads a package called \pkg{wpp}\eqn{x} where \eqn{x} is the \code{wpp.year} and generates 
historical migration rates using the 
\code{\link[wpp2019]{migration}} and \code{\link[wpp2019]{pop}} datasets.}

\item{my.mig.file}{File name containing user-specified historical time series of migration rates 
for all locations that should be included in the simulation.}

\item{sigma.c.min, a.up, a.ini, a.half.width, mu.range, sigma.mu.range, mu.ini}{Settings for the parameters
of the model (see Azose & Raftery 2015), such as minimum value, truncation ranges, slicing half width and initial values.
Initial values (*.ini) can be given as a vector of length \code{nr.chains}, giving one initial value per chain.
By default the initial values are equidistantly spread between their respective ranges.}

\item{exclude.from.world}{Vector of country codes that should not influence the hyperparameters. 
However, country-specific parameters will be generated for these countries.}

\item{seed}{Seed of the random number generator. If \code{NULL} no seed is set. It can be used to generate reproducible results.}

\item{parallel}{(NOT CURRENTLY IMPLEMENTED) Whether to run code in parallel, if available.}

\item{nr.nodes}{Relevant only if \code{parallel} is \code{TRUE}. It gives the number of nodes for running the simulation in parallel. 
By default it equals to the number of chains.}

\item{buffer.size}{Buffer size (in number of iterations) for keeping data in the memory before flushing to disk.}

\item{verbose}{Whether or not to print status updates to console window while code is running.}

\item{verbose.iter}{If verbose is TRUE, the number of iterations to wait between printing updates.}

\item{...}{Other arguments passed to \code{\link{run.mig.mcmc}}}
}
\value{
An object of class \code{bayesMig.mcmc.set} containing the sampled posterior parameter values
}
\description{
Runs MCMCs for simulating the net migration rate of all countries of the
world, using a Bayesian hierarchical model.
}
\examples{
run.mig.mcmc(nr.chains=2, iter=30, thin=1)
}
