% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outputs.R
\name{summary.bayesMig.convergence}
\alias{summary.bayesMig.convergence}
\title{Summary of Convergence Diagnostics}
\usage{
\method{summary}{bayesMig.convergence}(object, expand = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{\link{bayesMig.prediction}}.}

\item{expand}{By default, the function does not show country-specific parameters 
for which there was no convergence (only country-independent parameters), 
if the status is \sQuote{red}. This argument can switch that option on.}

\item{\dots}{Not used.}
}
\value{
List with items that summarize an object of class \code{\link{bayesMig.convergence}}.
}
\description{
Summary of an object of class \code{\link{bayesMig.convergence}} created 
    using the \code{\link{mig.diagnose}} function. It gives an overview about parameters 
    that did not converge.
}
