% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outputs.R
\name{get.mig.convergence.all}
\alias{get.mig.convergence.all}
\alias{get.mig.convergence}
\title{Accessing Convergence Diagnostics Object}
\usage{
get.mig.convergence.all(sim.dir = file.path(getwd(), "bayesMig.output"))

get.mig.convergence(
  sim.dir = file.path(getwd(), "bayesMig.output"),
  thin = 225,
  burnin = 10000
)
}
\arguments{
\item{sim.dir}{Simulation directory used for computing the diagnostics.}

\item{thin}{Thinning interval used with this diagnostics.}

\item{burnin}{Burn-in used for computing the diagnostics.}
}
\value{
\code{get.mig.convergence} returns an object of class \code{\link{bayesMig.convergence}}; \cr
    \code{get.mig.convergence.all} returns a list of objects of class \code{\link{bayesMig.convergence}}.
}
\description{
The function retrieves results of convergence diagnostics 
    (object of class \code{\link{bayesMig.convergence}} created by \code{\link{mig.diagnose}}) 
    from disk.
}
\details{
Function \code{get.mig.convergence} loads an object of class 
    \code{\link{bayesMig.convergence}} for the specific \code{thin} and \code{burnin}. 
    Function \code{get.mig.convergence.all} loads all \code{\link{bayesMig.convergence}} objects 
    available in \code{sim.dir}.
}
\seealso{
\code{\link{mig.diagnose}}, \code{\link{summary.bayesMig.convergence}}
}
