% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_storage.R, R/mig_adjust.R
\name{store.bayesMig.convergence}
\alias{store.bayesMig.convergence}
\alias{store.bayesMig.prediction}
\alias{get.mig.shift}
\title{Internal Functions and datasets of bayesMig}
\usage{
store.bayesMig.convergence(diag, thin, burnin, output.dir)

store.bayesMig.prediction(pred, output.dir = NULL)

get.mig.shift(country.code, pred)
}
\value{
None
}
\description{
These functions and datasets are not to be used directly by the user.
}
\keyword{internal}
