% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mig.R
\name{mig.predict}
\alias{mig.predict}
\alias{bayesMig.prediction}
\title{Generate posterior trajectories of net migration rates}
\usage{
mig.predict(
  mcmc.set = NULL,
  end.year = 2100,
  sim.dir = file.path(getwd(), "bayesMig.output"),
  replace.output = FALSE,
  start.year = NULL,
  nr.traj = NULL,
  thin = NULL,
  burnin = 20000,
  use.cummulative.threshold = FALSE,
  ignore.gcc.in.threshold = FALSE,
  ignore.last.observed = FALSE,
  save.as.ascii = 0,
  output.dir = NULL,
  seed = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{mcmc.set}{Object of class \code{bayesMig.mcmc.set} corresponding to sampled
parameter values for net migration model. If it is \code{NULL}, the object
is loaded from the directory specified in \code{sim.dir}}

\item{end.year}{End year of the prediction}

\item{sim.dir}{Directory with MCMC simulation results. It should be the same as
the \code{output.dir} argument in \code{\link{run.mig.mcmc}}}

\item{replace.output}{Logical value. If \code{TRUE}, existing predictions in
\code{output.dir} will be replaced by results of this run.}

\item{start.year}{Start year of the prediction, i.e. the first predicted year. By default the prediction is 
started at the next time period after \code{present.year} set in the estimation
step. If \code{start.year} is smaller than the default, projections for locations
and time periods that have data available after \code{start.year} are set to those data.}

\item{nr.traj}{Number of trajectories to be generated. 
If \code{NULL}, the argument \code{thin} is taken to determine the number of 
trajectories. If both are \code{NULL}, the number of trajectories
corresponds to the size of the parameter sample.}

\item{thin}{Thinning interval used for determining the number of trajectories. 
Only relevant if \code{nr.traj} is \code{NULL}.}

\item{burnin}{Number of iterations to be discarded from the beginning of the parameter traces.}

\item{use.cummulative.threshold}{If \code{TRUE} a historical cummulative threshold is applied 
derived from all rates, while for non-GCC countries the GCC countries are excluded. 
Six time periods are used in a 5-year simulation, corresponding to
30 years in an annual simulation. If this option is used in a non-country simulation,
e.g. in a sub-national settings, set the \code{ignore.gcc.in.threshold} argument to \code{TRUE}.}

\item{ignore.gcc.in.threshold}{If \code{use.cummulative.threshold} is \code{TRUE}, by default the GCC countries
(plus Western Sahara and Djibouti) identified by numerical codes of the countries are excluded from computing 
the historical cummulative thresholds. If this argument is \code{TRUE}, this distinction is not made.}

\item{ignore.last.observed}{By default, the prediction (or imputation) for each location starts 
one time period after the last observed data point for that location, defined by the 
\dQuote{last.observed} column in the data captured during the estimation. 
If this argument is set to \code{TRUE}, the prediction ignores the \dQuote{last.observed} 
value and starts at the last data point found in the data. This allows to exclude 
some time periods from the estimation (using \dQuote{last.observed}), but include them 
in the prediction.}

\item{save.as.ascii}{Either a number determining how many trajectories should be
converted into an ASCII file, or 'all' in which case all trajectories are converted.
It should be set to 0 if no conversion is desired.}

\item{output.dir}{Directory into which the resulting prediction object and the 
trajectories are stored. If it is \code{NULL}, it is set to either \code{sim.dir},
or to \code{output.dir} of \code{mcmc.set$meta} if \code{mcmc.set} is given.}

\item{seed}{Seed of the random number generator. If \code{NULL} no seed is set. 
Can be used to generate reproducible projections.}

\item{verbose}{Logical value. Switches log messages on and off.}

\item{...}{Other arguments passed to \code{\link{mig.predict}}}
}
\value{
A list with 11 components. Key result component is an array of \code{quantiles} with dimensions
(number of locations) x (number of computed quantiles) x (number of projected time points).
First time point in the sequence is not a projection, but the last observed time period.

Other key result components include \code{traj.mean.sd}, a summary of means and standard deviations for each country
at each time point. See \code{\link[bayesTFR]{bayesTFR.prediction}} for more detail.
}
\description{
Using the posterior parameter samples simulated by \code{\link{run.mig.mcmc}},
generate posterior trajectories for the net migration rates for all countries of
the world, or all locations included in the estimation. This code \emph{does not} adjust trajectories to ensure that net
migration counts sum to zero.
}
\details{
The trajectories of net migration rates for each location are generated using the model of Azose & Raftery (2015).
Parameter samples  simulated via \code{\link{run.mig.mcmc}} are used from all chains, from which the given \code{burnin} 
was discarded. They are evenly thinned to match \code{nr.traj} or using the \code{thin} argument. 
Such thinned parameter traces, collapsed into one chain, if they do not already exist, are stored on disk 
into the sub-directory \file{thinned_mcmc_\emph{t}_\emph{b}} where \emph{t} is the value  of \code{thin} and 
\emph{b} the value of \code{burnin}.

The projection is run for all missing values before the present year, if any. 
Medians over the trajectories are used as imputed values and the trajectories are discarded. 
The process then continues by projecting the future values where all generated trajectories are kept.

A special case is when the argument \code{start.year} is given that is smaller than or equal to
the present year. In such a case, imputed missing values before present year are treated 
as ordinary predictions (trajectories are kept). All historical data between start year 
and present year are used as projections.

The resulting prediction object is saved into \file{\{output.dir\}/predictions}. Trajectories 
for all locations are saved into the same directory in a binary format, one file per location. 
At the end of the projection, if \code{save.as.ascii} is larger than 0, the function converts 
the given number of trajectories into a CSV file. They are selected by equal spacing. 
In addition, two summary files are created: one in a user-friendly format, the other using 
a UN-specific coding.
}
\examples{
# Toy simulation for US states
us.mig.file <- file.path(find.package("bayesMig"), "extdata", "USmigrates.txt")
sim.dir <- tempfile()
m <- run.mig.mcmc(nr.chains = 2, iter = 30, thin = 1, my.mig.file = us.mig.file, 
        output.dir = sim.dir, present.year = 2017, annual = TRUE)

# Prediction
pred <- mig.predict(sim.dir = sim.dir, burnin = 5, end.year = 2050)
# here unrealistic results since this is a toy simulation 
mig.trajectories.plot(pred, "Hawaii", pi = 80, ylim = c(-0.02, 0.02)) 
mig.trajectories.table(pred, "Hawaii")

unlink(sim.dir, recursive = TRUE)
# For projections on national level, see ?bayesMig.
}
