% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcmc.R
\name{run.mig.mcmc}
\alias{run.mig.mcmc}
\alias{bayesMig.mcmc.set,}
\alias{bayesMig.mcmc,}
\alias{bayesMig.mcmc.meta}
\title{Run Markov chain Monte Carlo for parameters of net migration rate model}
\usage{
run.mig.mcmc(
  nr.chains = 3,
  iter = 50000,
  output.dir = file.path(getwd(), "bayesMig.output"),
  thin = 1,
  replace.output = FALSE,
  annual = FALSE,
  start.year = 1950,
  present.year = 2020,
  wpp.year = 2019,
  my.mig.file = NULL,
  sigma.c.min = 1e-04,
  a.up = 10,
  a.ini = NULL,
  a.half.width = 0.3,
  mu.range = c(-0.5, 0.5),
  sigma.mu.range = c(0, 0.5),
  mu.ini = NULL,
  exclude.from.world = NULL,
  seed = NULL,
  parallel = FALSE,
  nr.nodes = nr.chains,
  buffer.size = 1000,
  verbose = TRUE,
  verbose.iter = 10,
  ...
)
}
\arguments{
\item{nr.chains}{An integer number of independent Markov chains to run.}

\item{iter}{The number of iterations to run per Markov chain.}

\item{output.dir}{A file path pointing to the directory in which to store results.}

\item{thin}{Thinning interval -- A chain with 1000 iterations thinned by 20 will return a 
final count of 50 iterations.}

\item{replace.output}{If the specified output directory already exists, should it be overwritten?}

\item{annual}{If \code{TRUE}, the model assumes the underlying data is on annual time scale. 
In such a case, argument \code{my.mig.file} must be used to provide the annual observed data.}

\item{start.year}{Start year for using historical data.}

\item{present.year}{End year for using historical data.}

\item{wpp.year}{Year for which WPP data is used if no user data is provided via \code{my.mig.file}. 
In such a case, the function loads a package called \pkg{wpp}\eqn{x} where \eqn{x} is the \code{wpp.year} and generates 
historical migration rates using the 
\code{\link[wpp2019]{migration}} and \code{\link[wpp2019]{pop}} datasets.}

\item{my.mig.file}{File name containing user-specified historical time series of migration rates 
for all locations that should be included in the simulation.}

\item{sigma.c.min, a.up, a.ini, a.half.width, mu.range, sigma.mu.range, mu.ini}{Settings for the parameters
of the model (see Azose & Raftery 2015), such as minimum value, truncation ranges, slicing half width and initial values.
Initial values (*.ini) can be given as a vector of length \code{nr.chains}, giving one initial value per chain.
By default the initial values are equidistantly spread between their respective ranges.}

\item{exclude.from.world}{Vector of country codes that should not influence the hyperparameters. 
However, country-specific parameters will be generated for these countries.}

\item{seed}{Seed of the random number generator. If \code{NULL} no seed is set. It can be used to generate reproducible results.}

\item{parallel}{Whether to run code in parallel.}

\item{nr.nodes}{Relevant only if \code{parallel} is \code{TRUE}. It gives the number of nodes for running the simulation in parallel. 
By default it equals to the number of chains.}

\item{buffer.size}{Buffer size (in number of iterations) for keeping data in the memory before flushing to disk.}

\item{verbose}{Whether or not to print status updates to console window while code is running.}

\item{verbose.iter}{If verbose is TRUE, the number of iterations to wait between printing updates.}

\item{\dots}{Additional parameters to be passed to the function \code{\link[snowFT]{performParallel}}, if \code{parallel} is \code{TRUE}.}
}
\value{

}
\description{
Runs MCMCs for simulating the net migration rate of all countries of the
world (or locations specified by users), using the Bayesian hierarchical model of Azose & Raftery (2015).
}
\details{
The function creates an object of class \code{\link{bayesMig.mcmc.meta}} and 
stores it in \code{output.dir}. It launches \code{nr.chains} MCMCs, either sequentially or 
in parallel. Parameter traces of each chain are stored as ASCII files in a subdirectory 
of \code{output.dir}, called \code{mc}\emph{x} where \emph{x} is the identifier of that chain. 
There is one file per parameter, named after the parameter with the suffix \dQuote{.txt}.
Country-specific parameters have the suffix \code{_country}\emph{c} where \emph{c} is the location code.
In addition to the trace files, each \code{mc}\emph{x} directory contains the object 
\code{\link{bayesMig.mcmc}} in binary format.  
All chain-specific files  are written onto disk after the first, last and each 
\eqn{i}-th (thinned) iteration, where \eqn{i} is given by the argument \code{buffer.size}.

By default (if no data is passed via the \code{my.mig.file} argument), the function 
loads observed data (further denoted as WPP dataset), from the \code{\link[wpp2019]{migration}} 
and \code{\link[wpp2019]{pop}} datasets in the \pkg{wpp}\eqn{x} package where \eqn{x} is 
the \code{wpp.year}. Net migration rates are computed as migration(\eqn{t}) / (population(\eqn{t_e}) - migration(\eqn{t})) 
where \eqn{t_e} means the end of time period \eqn{t}. 

The argument \code{my.mig.file} can be used to overwrite the default data. 
If it is used, it should contain the net migration rate for all locations to be used in the simulation, as no WPP data is used 
in such a case. The structure of the file has the same format as the \code{\link[wpp2019]{migration}} dataset,
but the values should be rates (instead of counts). Each row corresponds to a location. It does not have 
to be necessarily a country - it can be for example a subnational unit. It must contain columns 
\dQuote{country_code} or \dQuote{code} (unique identifier of the location), \dQuote{name}, and columns representing 
5-year time intervals (if \code{annual} is \code{FALSE}), e.g., \dQuote{1995-2000}, \dQuote{2000-2005} etc. 

If \code{annual} is \code{TRUE} the default WPP dataset is not used and the \code{my.mig.file} argument 
must provide the dataset to be used for estimation. Its time-related columns should be single years.
An example dataset of annual net migration rates for US states is included in the package, see example below. 

If there are countries or locations that should be excluded from influencing the hyperparameters,
for example small countries or locations with unique migration patterns, their codes 
should be included in the argument \code{exclude.from.world}. These locations will still get 
their parameters simulated and thus, can be included in a projection.
}
\examples{
\dontrun{
# Toy simulation for US states
us.mig.file <- file.path(find.package("bayesMig"), "extdata", "USmigrates.txt")
sim.dir <- tempfile()
m <- run.mig.mcmc(nr.chains = 2, iter = 30, thin = 1, my.mig.file = us.mig.file, 
        annual = TRUE, output.dir = sim.dir)
summary(m)
summary(m, "Washington")

# later one can access the object from disk
m <- get.mig.mcmc(sim.dir)
 
unlink(sim.dir, recursive = TRUE)
# For a country-level simulation, see example in ?bayesMig. 
}
}
\references{
Azose, J. J., & Raftery, A. E. (2015). 
Bayesian probabilistic projection of international migration. Demography, 52(5), 1627-1650.
}
\seealso{
\code{\link{get.mig.mcmc}}, \code{\link{summary.bayesMig.mcmc.set}}, \code{\link{mig.partraces.plot}},
\code{\link{mig.pardensity.plot}}, \code{\link{mig.predict}}
}
