% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcmc.R
\name{run.mig.mcmc}
\alias{run.mig.mcmc}
\title{Run Markov chain Monte Carlo for parameters of net migration rate model}
\usage{
run.mig.mcmc(nr.chains = 3, iter = 1000, thin = 1, verbose = TRUE,
  verbose.iter = 10, output.dir = file.path(getwd(), "bayesMig.output"),
  replace.output = FALSE, parallel = FALSE, ...)
}
\arguments{
\item{nr.chains}{An integer number of independent Markov chains to run.}

\item{iter}{The number of iterations to run per Markov chain.}

\item{thin}{Thinning interval -- A chain with 1000 iterations thinned by 20 will return a 
final count of 50 iterations.}

\item{verbose}{Whether or not to print status updates to console window while code is running.}

\item{verbose.iter}{If verbose is TRUE, the number of iterations to wait between printing updates.}

\item{output.dir}{A file path pointing to the directory in which to store results.}

\item{replace.output}{If the specified output directory already exists, should it be overwritten?}

\item{parallel}{(NOT CURRENTLY IMPLEMENTED) Whether to run code in parallel, if available.}

\item{...}{Other arguments passed to \code{\link{run.mig.mcmc}}}
}
\value{
An object of class \code{bayesMig.mcmc.set} containing the sampled posterior parameter values
}
\description{
Runs (or continues running) MCMCs for simulating the net migration rate of all countries of the
world, using a Bayesian hierarchical model.
}
\examples{
run.mig.mcmc(nr.chains=2, iter=30, thin=1)
}
